Chat testing (#50)
* fixed chat issues + added some testing in clientSocket for chat

* Fixed frontend build issues, and npm audited vulnurabilities on frontend and backend
