+++ backend/package.json
@@ -16,6 +16,7 @@
   "license": "ISC",
   "description": "",
   "dependencies": {
+    "brace-expansion": "^4.0.1",
     "concurrently": "^9.1.2",
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
@@ -26,7 +27,8 @@
     "socket.io": "^4.8.1",
     "socket.io-client": "^4.8.1",
     "typescript": "^5.7.3",
-    "vitest": "^3.0.4"
+    "vite": "^7.1.10",
+    "vitest": "^3.2.4"
   },
   "devDependencies": {
     "tsc-alias": "^1.8.16",
