+++ frontend/src/features/Chat/ChatWidgetActiveChat.tsx
@@ -111,6 +111,8 @@ export default function ChatWidgetActiveChat({
     userIDsInChat[0] !== self.id ? userIDsInChat[0] : userIDsInChat[1];
   const otherUserObj = users[otherUserID];
 
+  console.log('activebru', activeChatObj);
+
   return (
     <>
       <div
