+++ shared/package.json
@@ -4,7 +4,7 @@
   "main": "index.js",
   "scripts": {
     "testAll": "vitest",
-    "test": "vitest clientSocket.mentorshipRequest.test.ts"
+    "test": "vitest clientSocket.chat.test.ts"
   },
   "keywords": [],
   "author": "",
