+++ shared/tsconfig.json
@@ -21,7 +21,5 @@
   // lets ts know where to look for files
   // note: files not included will not be compiled, nor will they be able to use aliases
   "include": ["."],
-  // exclude shared's vitest config to avoid build issues
-  "exclude": ["../shared/vitest.config.ts", "../shared/**/*.test.ts", "../shared/**/vitestSetup.ts"], 
   "compileOnSave": true
 }
