+++ shared/validation/general.ts
@@ -485,3 +485,20 @@ export function isAssessment(s: object): s is Assessment {
   return true;
 }
 
+
+/**
+ * Checks s is valid message.
+ *
+ * Throws descriptive error if invalid. Returns true otherwise.
+ * @param s
+ * @returns
+ */
+export function isValidMessageContent(s: unknown): s is string {
+  if (!s || typeof s != "string") {
+    throw new Error("No message content was provided");
+  }
+  if (s.trim().length == 0) {
+    throw new Error("Message content is too short");
+  }
+  return true;
+}
